[Bindings] Simplify [RequiresExistingAtomicString] IDL extended attribute handling https://bugs.webkit.org/show_bug.cgi?id=157465
Reviewed by Darin Adler.

Source/WebCore:

Simplify [RequiresExistingAtomicString] IDL extended attribute handling
in the bindings generator.

* bindings/scripts/CodeGeneratorJS.pm:
(GenerateParametersCheck):
Drop code specific to [RequiresExistingAtomicString].

(JSValueToNative):
Deal with [RequiresExistingAtomicString] IDL extended attribute,
similarly to [AtomicString] and [TreatNullAs].

* bindings/scripts/test/JS/JSTestObj.cpp:
* bindings/scripts/test/TestObj.idl:
Add bindings test coverage.

* dom/DocumentFragment.cpp:
(WebCore::DocumentFragment::getElementById):
* svg/SVGSVGElement.cpp:
(WebCore::SVGSVGElement::getElementById):
Add null check for the id parameter. The generated bindings used to do
this null check for us but it is no longer the case (to make the bindings
generator a bit simpler). This extended attribute is not commonly used
and is meant as a micro-optimization for getElementById().

Source/WTF:

Add an AtomicString(RefPtr<AtomicStringImpl>&&) constructor that is now
used by the bindings when [RequiresExistingAtomicString] is used. This
is more efficient than using AtomicString(AtomicStringImpl*) as the
caller has a RefPtr<AtomicStringImpl> and is consistent with the
pre-existing String(RefPtr<AtomicStringImpl>&&) constructor.

* wtf/text/AtomicString.h:
(WTF::AtomicString::AtomicString):

Identifier: 175593@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200562 268f45cc-cd09-0410-ab3c-d52691b4dbfc