Web Inspector: CRASH getting internal properties of function with no bound arguments causes https://bugs.webkit.org/show_bug.cgi?id=157613 <rdar://problem/26238754>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-05-12
Reviewed by Timothy Hatcher.

Source/JavaScriptCore:

* inspector/JSInjectedScriptHost.cpp:
(Inspector::JSInjectedScriptHost::getInternalProperties):
Gracefully handle a JSBoundFunction with no bound arguments.
In this case boundArgs is JSValue() which we don't want to
expose as the value of the internal property.

LayoutTests:

* inspector/runtime/getProperties-expected.txt:
* inspector/runtime/getProperties.html:
Include coverage for a bound function with no bound arguments.

Identifier: 175747@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200746 268f45cc-cd09-0410-ab3c-d52691b4dbfc