[JSC] Move the CheckTierUp function calls out of the main path https://bugs.webkit.org/show_bug.cgi?id=157668
Patch by Benjamin Poulain <bpoulain@apple.com> on 2016-05-13
Reviewed by Mark Lam.

If you have a tiny tiny loop (for example, Sunspider's bits-in-byte),
the size of CheckTierUp is a problem.

On multi-issue CPUs, the node is so big that we do not
get to run anything from the loop in the instruction fetch.

On x86, having a bigger loop also pushes us out of the LSD.

This is a 6% improvement on bits-in-byte. Other Sunspider tests
only improves marginally.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::addSlowPathGenerator):
(JSC::DFG::SpeculativeJIT::runSlowPathGenerators):
* dfg/DFGSpeculativeJIT.h:
(JSC::DFG::SpeculativeJIT::silentSpill):
(JSC::DFG::SpeculativeJIT::silentFill):
* dfg/DFGSpeculativeJIT64.cpp:
(JSC::DFG::SpeculativeJIT::compile):


Identifier: 175825@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200897 268f45cc-cd09-0410-ab3c-d52691b4dbfc