[JSC] SetLocal without exit do not need phantoms https://bugs.webkit.org/show_bug.cgi?id=157653
Patch by Benjamin Poulain <bpoulain@apple.com> on 2016-05-13
Reviewed by Filip Pizlo.

I made a mistake in r200498.

If a SetLocal cannot possibly exit, we were not clearing
the source of the operand. As a result, we sometime kept
a value alive up to the end of the block.

That's uncommon because SetLocal typically appear
toward the end of blocks. That's probably why there was
no perf impact with that fix.

* dfg/DFGPhantomInsertionPhase.cpp:


Identifier: 175826@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200898 268f45cc-cd09-0410-ab3c-d52691b4dbfc