TypeSet/StructureShape have a flawed sense of JS prototype chains https://bugs.webkit.org/show_bug.cgi?id=157760
Reviewed by Joseph Pecoraro.

There was an assumption that we would bottom out in "Object". This is
not true for many reasons. JS objects may not end in Object.prototype.
Also, our mechanism of grabbing an Object's class name may also not
bottom out in "Object". We were seeing this in the JS objects we use
in the InjectedScriptSource.js inspector script.

* runtime/TypeSet.cpp:
(JSC::StructureShape::leastCommonAncestor):
* tests/typeProfiler/weird-prototype-chain.js: Added.
(wrapper.foo):
(wrapper.let.o2):
(wrapper):


Identifier: 175901@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200980 268f45cc-cd09-0410-ab3c-d52691b4dbfc