REGRESSION: JetStream crashes on some iPhones https://bugs.webkit.org/show_bug.cgi?id=157814
Reviewed by Michael Saboff.

* bmalloc/Sizes.h: Reduce smallMax to 32kB.

Previous justification for 64kB was:

    * bmalloc/Sizes.h: Upped smallMax to 64kB. Upping to 32kB is pretty
    reasonable, since sizes between 16kB and 32kB share page sizes. I went
    all the way up to 64kB because the GC uses 64kB blocks, and also just
    for extra padding to ensure that large allocations are indeed rare.

It turns out that the bump to 64kB substantially increases our memory
high water mark on JetStream, leading to jetsam crashes. Also, there
doesn't seem to be a practical performance problem to putting objects in
the (32kB - 64kB) range in the large allocator.


Identifier: 175938@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201050 268f45cc-cd09-0410-ab3c-d52691b4dbfc