[GStreamer] webaudio playback improvements https://bugs.webkit.org/show_bug.cgi?id=155228
Reviewed by Michael Catanzaro.

* platform/audio/gstreamer/AudioDestinationGStreamer.cpp:
(WebCore::autoAudioSinkChildAddedCallback): Fix sink buffer-time
to 100ms to reduce playback latency.
(WebCore::AudioDestinationGStreamer::AudioDestinationGStreamer):
Connect to child-added signal of autoaudiosink to be notified when
a real sink is added into the bin.
* platform/audio/gstreamer/WebKitWebAudioSourceGStreamer.cpp:
(webKitWebAudioSrcConstructed): Fine-tune blocksize of appsrc
according to the buffer size already configured on the src element.

Identifier: 175958@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201077 268f45cc-cd09-0410-ab3c-d52691b4dbfc