r199812 broke test262 https://bugs.webkit.org/show_bug.cgi?id=157595
Reviewed by Filip Pizlo.

Source/JavaScriptCore:

Added a reasonable limit to the size of the match result array to catch possible
infinite loops when matching.
Added a new tests that creates an infinite loop in RegExp.prototype.[Symbol.match]
by creating a subclass of RegExp where the base RegExp's global flag is false and
the subclass overrides .global with a getter that always returns true.

* builtins/RegExpPrototype.js:
(match):
* tests/stress/regress-157595.js: Added.
(MyRegExp):
(MyRegExp.prototype.get global):
(test):
(catch):

Tools:

Added a new run type, runOneLargeHeap, for tests that use a large amount of memory.
This run type will not run with the --memory-limited option.  Without that option,
we'll only the default test variant.

* Scripts/run-jsc-stress-tests:


Identifier: 175984@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201105 268f45cc-cd09-0410-ab3c-d52691b4dbfc