Unreviewed test fix after r201427. https://bugs.webkit.org/show_bug.cgi?id=157423 <rdar://problem/23751632>
A debug assertion was firing during some test runs due to the Geolocation permission
being turned off during the test. The timer logic was originally written to assert
if the timer fired when permissions were disabled. But this is no longer valid,
because we expect the Geolocation system to be active and become deactivated if the
browsing context violates one of the security criteria.

* DumpRenderTree/mac/UIDelegate.mm:
(-[UIDelegate timerFired]): Remove invalid assertion.



Identifier: 176237@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201434 268f45cc-cd09-0410-ab3c-d52691b4dbfc