AX: [ATK] accessibility/gtk/no-notification-for-unrendered-iframe-children.html began failing after r201416 https://bugs.webkit.org/show_bug.cgi?id=158152
Reviewed by Chris Fleizach.

The failure is actually a bug fix because only one child is being added, but two
notifications were being emitted. Now there is only one notification. To verify
this was the case, we really should examine the child reportedly being added. That
child is the ATK event's any_data, so pass along that child to the listener.

Tools:

* WebKitTestRunner/InjectedBundle/atk/AccessibilityNotificationHandlerAtk.cpp:

LayoutTests:

The failing test and associated expectations were modified to remove the duplicate
notification and include the title of the added child for the notification we get.

* accessibility/gtk/no-notification-for-unrendered-iframe-children.html: Updated.
* accessibility/gtk/no-notification-for-unrendered-iframe-children-expected.txt: Updated.


Identifier: 176261@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201458 268f45cc-cd09-0410-ab3c-d52691b4dbfc